+++ app/lib/shared/versions.dart
@@ -21,9 +21,9 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// Make sure that at least two versions are kept here as the next candidates
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
-  '2021.04.06', // The current [runtimeVersion].
+  '2021.04.19', // The current [runtimeVersion].
+  '2021.04.06',
   '2021.03.19',
-  '2021.03.15',
 ];
 
 /// Represents a combined version of the overall toolchain and processing,
@@ -53,10 +53,10 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '2.12.0';
-final String toolStableDartSdkVersion = '2.12.2';
-final String toolStableFlutterSdkVersion = '2.0.4';
-final String toolPreviewDartSdkVersion = '2.13.0-116.1.beta';
-final String toolPreviewFlutterSdkVersion = '2.1.0-12.2.pre';
+final String toolStableDartSdkVersion = '2.12.4';
+final String toolStableFlutterSdkVersion = '2.0.5';
+final String toolPreviewDartSdkVersion = '2.13.0-211.6.beta';
+final String toolPreviewFlutterSdkVersion = '2.2.0-10.1.pre';
 
 // Value comes from package:pana.
 final String panaVersion = pana.packageVersion;